moved plot handling methods into new PlotManager class, gives some keyerror although seems to run normally
