+++ src/managers/__init__.py
@@ -1,3 +1,4 @@
 from .data_holder import DataHolder
 from .timer_service import TimerService
 from .device_manager import DeviceManager
+from .plot_manager import PlotManager
